`close-radius` prolog predicate for objects to grasp
When simulating, this predicate forces the PR2 to completely open its gripper when grasping. This is due to the fact that we don't actually, physically grasp in Gazebo, but use the capabilities of the `attache` plugin to create a virtual joint between the robot's gripper and the object to grasp while carrying it.